+++ dacd/src/app/app.module.ts
@@ -23,6 +23,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatIconModule } from '@angular/material/icon';
+import { FavoritesComponent } from './components/favorites/favorites.component';
 
 @NgModule({
   declarations: [
@@ -30,7 +31,8 @@ import { MatIconModule } from '@angular/material/icon';
     HomepageComponent,
     HeaderComponent,
     FooterComponent,
-    UserContentComponent
+    UserContentComponent,
+    FavoritesComponent
   ],
   imports: [
     BrowserModule,