+++ CHANGELOG.md
@@ -4,14 +4,18 @@ Versions history changes register
 
 ## Version History
 
+* 1.2.0
+  * ADD: New component `favorites` and use `@Angular/core/Output`
+  * CHANGED: Refatoring homepage and components
+
 * 1.1.1
   * FIX: Favorites show if null
 
 * 1.1.0
-  * ADD: Favorites feature
+  * ADD: `Favorites` feature
 
 * 1.0.1
-  * ADD: README informations of project
+  * ADD: `README` informations of project
 
 * 1.0.0
   * Project conclueded